requirements.txt
pyyaml==6.0.2
langchain==0.3.0
langchain-community==0.3.0
openai==1.44.0
nltk==3.9.1
tqdm==4.67.1
networkx==3.2.1
pandas==2.2.3
numpy
aiofiles
langchain_openai==0.2.1
faiss-cpu